requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
openai
python-dotenv
alembic
requests
PyYAML
